+++ libs/f-flow/src/f-connection-v2/utils/connection-line-builder/builders/utils/index.ts
@@ -2,6 +2,7 @@ export * from './build-connection-anchors';
 export * from './calculate-center-between-points';
 export * from './calculate-curve-candidates';
 export * from './calculate-polyline-candidates';
+export * from './calculate-smooth-control-point';
 export * from './merge-point-chains';
 export * from './multi-cubic';
 export * from './normalize-polyline';